Rename Table components to DictionaryTable and add new "add to dictionary" route
