+++ components.d.ts
@@ -10,12 +10,15 @@ declare module 'vue' {
     AddStudentModal: typeof import('./src/components/teachers/AddStudentModal.vue')['default']
     AddTagModal: typeof import('./src/components/table/AddTagModal.vue')['default']
     AddToStudentDictionary: typeof import('./src/components/students/AddToStudentDictionary.vue')['default']
+    AddWordForm: typeof import('./src/components/dictionary/AddWordForm.vue')['default']
     Button: typeof import('primevue/button')['default']
     Column: typeof import('primevue/column')['default']
     DataTable: typeof import('primevue/datatable')['default']
     DeleteStudentModal: typeof import('./src/components/teachers/DeleteStudentModal.vue')['default']
     DeleteWordModal: typeof import('./src/components/students/DeleteWordModal.vue')['default']
     Dialog: typeof import('primevue/dialog')['default']
+    DictionaryTable: typeof import('./src/components/table/DictionaryTable.vue')['default']
+    DictionaryTableContainer: typeof import('./src/components/table/DictionaryTableContainer.vue')['default']
     Dropdown: typeof import('primevue/dropdown')['default']
     EditNotesModal: typeof import('./src/components/students/EditNotesModal.vue')['default']
     EditStudentModal: typeof import('./src/components/teachers/EditStudentModal.vue')['default']
@@ -40,8 +43,6 @@ declare module 'vue' {
     Signup: typeof import('./src/components/auth/Signup.vue')['default']
     StudentHeader: typeof import('./src/components/students/StudentHeader.vue')['default']
     StudentWordTable: typeof import('./src/components/students/StudentWordTable.vue')['default']
-    Table: typeof import('./src/components/table/Table.vue')['default']
-    TableContainer: typeof import('./src/components/table/TableContainer.vue')['default']
     Tag: typeof import('primevue/tag')['default']
     TeacherTable: typeof import('./src/components/teachers/TeacherTable.vue')['default']
     Textarea: typeof import('primevue/textarea')['default']
