+++ src/components/table/DictionaryTable.vue
