+++ src/router/index.ts
@@ -1,18 +1,20 @@
 import { checkAuth } from '@/router/middleware'
-import HomeView from '@/views/HomeView.vue'
+
 import LoginView from '@/views/LoginView.vue'
 
 import StudentView from '@/views/student/StudentView.vue'
 import TeacherView from '@/views/TeacherView.vue'
+import DictionaryView from '@/views/DictionaryView.vue'
+import DictionaryEntryView from '@/views/DictionaryEntryView.vue'
 import { createRouter, createWebHistory } from 'vue-router'
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
   routes: [
     {
-      path: '/home',
-      name: 'Home',
-      component: HomeView
+      path: '/home/dictionary',
+      name: 'Dictionary',
+      component: DictionaryView
     },
     {
       path: '/',
@@ -37,6 +39,11 @@ const router = createRouter({
       path: '/home/teacher',
       name: 'Teacher',
       component: TeacherView
+    },
+    {
+      path: '/home/dictionary-entry',
+      name: 'DictionaryEntry',
+      component: DictionaryEntryView
     }
   ]
 })
