+++ src/utils/pocketbaseConnection.ts
@@ -1,4 +1,3 @@
 import Pocketbase from 'pocketbase'
 
-// TODO Absolutely cannot forget to change this if I make the repo public
 export const pb = new Pocketbase('https://pocket.jaffs.org')
