+++ src/views/TeacherView.vue
@@ -3,5 +3,3 @@
     <TeacherTable />
   </div>
 </template>
-<script setup lang="ts"></script>
-<style lang="scss" scoped></style>
